agents/kiln_agent/requirements.txt
flask==2.3.3
gym==0.26.2
numpy==1.24.4
# Google Cloud libraries
google-cloud-pubsub==2.19.4
google-cloud-monitoring==2.14.1
